requirements.txt
langchain
langchain-community
langchain-openai
psycopg[binary]
python-dotenv
langchain-postgres
